Some refactoring to fix VC++ warnings

+++ first_ray/camera.h
@@ -19,7 +19,7 @@ struct camera
     camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov, float aspect, float aperture, float focus_dist)
     {
         lens_radius = aperture / 2;
-        float theta = vfov * M_PI / 180;
+        float theta = vfov * (float)M_PI / 180.0f;
         float half_height = tan(theta / 2);
         float half_width = aspect * half_height;
