+++ first_ray/hitable_list.h
@@ -22,7 +22,7 @@ bool hitable_list::hit(const ray &r, float t_min, float t_max, hit_record &rec) 
 
     for (int i = 0; i < list_size; i++)
     {
-        if (list[i]->hit(r, t_min, closest_so_far, temp_rec))
+        if (list[i]->hit(r, t_min, (float)closest_so_far, temp_rec))
         {
             hit_anything = true;
             closest_so_far = temp_rec.t;
